add countries and airports to data

+++ src/SkyExplorer.Common/Models/BoundingBox.cs
@@ -2,8 +2,8 @@ namespace SkyExplorer.Common.Models;
 
 public class BoundingBox
 {
-    public double MinLatitude { get; set; }
-    public double MaxLatitude { get; set; }
-    public double MinLongitude { get; set; }
-    public double MaxLongitude { get; set; }
+    public decimal MinLatitude { get; set; }
+    public decimal MaxLatitude { get; set; }
+    public decimal MinLongitude { get; set; }
+    public decimal MaxLongitude { get; set; }
 }
